+++ src/ska_integration_test_harness/facades/tmc_facade.py
@@ -38,7 +38,7 @@ from ..structure.telescope_wrapper import TelescopeWrapper
 class TMCFacade:
     """A facade to TMC devices and the actions on them.
 
-    A facade to TMC sub-system, providing a simplified interface to:
+    A facade to TMC subsystem, providing a simplified interface to:
 
     - references to central node device,
     - references to subarray devices and their leaf devices,
@@ -152,7 +152,7 @@ class TMCFacade:
     ) -> Tuple[ResultCode, str]:
         """Invoke LoadDishCfg command on central Node.
 
-        :param dish_vcc_config: Dish vcc configuration json string.
+        :param dish_vcc_config: dish vcc configuration json string.
         :param wait_termination: set to False if you don't want to
             wait for the termination condition. By default the termination
             condition is waited.