+++ src/ska_integration_test_harness/facades/tmc_subarray_node_facade.py
@@ -59,7 +59,7 @@ from ska_integration_test_harness.structure.telescope_wrapper import (  # pylint
 class TMCSubarrayNodeFacade:
     """A facade to TMC Subarray Node device and its actions.
 
-    A facade to TMC sub-system, providing a simplified interface to the
+    A facade to TMC subsystem, providing a simplified interface to the
     subarray node devices and their actions. It contains:
 
     - references to subarray node device,
@@ -72,8 +72,7 @@ class TMCSubarrayNodeFacade:
     - various other actions (e.g., five point calibration scan).
 
     **IMPORTANT NOTE**: This facade is deprecated,
-    since we want to use a facade for each
-    sub-system. Please use
+    since we want to use a facade for each subsystem. Please use
     :py:class:`~ska_integration_test_harness.facades.tmc_facade.TMCFacade`
     instead.
     """
@@ -93,7 +92,7 @@ class TMCSubarrayNodeFacade:
 
     @property
     def dish_leaf_node_list(self):
-        """Return Dish Leaf Node List"""
+        """Return the dish leaf nodes as a list of device proxies."""
         return self._telescope.tmc.dish_leaf_node_list
         # NOTE: in old test harness code, sometimes just the first two
         # dishes were used. I don't know if there was a reason behind that
@@ -135,7 +134,7 @@ class TMCSubarrayNodeFacade:
         self._telescope.set_subarray_id(requested_subarray_id)
 
     # -----------------------------------------------------------
-    # Actions over subarray telescope state
+    # Actions on subarray telescope state
 
     def move_to_on(self, wait_termination: bool = True):
         """Move subarray to ON state.
@@ -164,7 +163,7 @@ class TMCSubarrayNodeFacade:
         return action.execute()
 
     # -----------------------------------------------------------
-    # Actions over subarray obs state
+    # Actions on subarray obs state
 
     # @sync_configure(device_dict=device_dict)
     def configure(