+++ src/ska_integration_test_harness/init/test_harness_builder.py
@@ -39,9 +39,9 @@ class TestHarnessBuilder:
     JSON inputs for various commands/purposes.
 
     The class uses also a set of tools to read and validate the configurations
-    and the inputs. Potentially, you can inject your own tools to customize
+    and the inputs. Potentially, you can inject your own tools to customise
     the behaviour of the builder. Potentially, you can also subclass this
-    builder to customize its behaviour.
+    builder to customise its behaviour.
 
     Usage example:
 
@@ -86,7 +86,7 @@ class TestHarnessBuilder:
         """The Kubernetes namespace where the SUT is deployed.
 
         It is optional, but if you set it, it will be used to connect to
-        get more information about the various subsystems devices.
+        get more information about the various subsystem devices.
         """
 
         # --------------------------------------------------------------