+++ src/ska_integration_test_harness/structure/__init__.py
@@ -3,14 +3,14 @@
 :py:class:`~ska_integration_test_harness.structure.telescope_wrapper.TelescopeWrapper`
 is the main entry point for the telescope system.
 
-Right now, it exposes the following sub-systems (through a wrapper for each):
+Right now, it exposes the following subsystems (through a wrapper for each):
 
 - TMC
 - SDP
 - CSP
 - Dishes
 
-All the subsystems wrappers are represented by abstract classes, so you can
+All the subsystem wrappers are represented by abstract classes, so you can
 create your own implementation for each one (e.g., an implementation that
 points to the devices of a real subsystem, or an implementation that points
 to the devices of an emulated subsystem).