+++ tests/init/test_test_harness_builder.py
@@ -26,10 +26,10 @@ from ska_integration_test_harness.structure.telescope_wrapper import (
 class TestTestHarnessBuilder:
     """Unit tests for TestHarnessBuilder.
 
-    Since this is a particularly critical component, we test it minimizing
+    Since this is a particularly critical component, we test it minimising
     the use of mocks and patches. We use them only when strictly necessary
     (e.g., Tango API). The used configurations are realistic ones, taken
-    from TMC-CSP MID integration tests. The passed input instead are JSON,
+    from TMC-CSP Mid integration tests. The passed input instead are JSON,
     but not semantically meaningful (it doesn't matter for the tests,
     since they will be needed only in teardown procedures or similar).
     """